Upgraded to grails 2.4.3 and spring-security-core:2.0-RC4
