+++ application.properties
@@ -1,5 +1,5 @@
 #Grails Metadata file
 #Sat Jul 06 10:14:33 CDT 2013
-app.grails.version=2.2.1
+app.grails.version=2.4.3
 app.name=smart-notes
-app.version=0.1
+app.version=0.2
