+++ grails-app/conf/BuildConfig.groovy
@@ -1,4 +1,4 @@
-grails.servlet.version = "2.5" // Change depending on target container compliance (2.5 or 3.0)
+grails.servlet.version = "3.0" // Change depending on target container compliance (2.5 or 3.0)
 grails.project.class.dir = "target/classes"
 grails.project.test.class.dir = "target/test-classes"
 grails.project.test.reports.dir = "target/test-reports"
@@ -6,65 +6,77 @@ grails.project.target.level = 1.6
 grails.project.source.level = 1.6
 //grails.project.war.file = "target/${appName}-${appVersion}.war"
 
-// uncomment (and adjust settings) to fork the JVM to isolate classpaths
-//grails.project.fork = [
-//   run: [maxMemory:1024, minMemory:64, debug:false, maxPerm:256]
-//]
+grails.project.fork = [
+  // configure settings for compilation JVM, note that if you alter the Groovy version forked compilation is required
+  //  compile: [maxMemory: 256, minMemory: 64, debug: false, maxPerm: 256, daemon:true],
 
+  // configure settings for the test-app JVM, uses the daemon by default
+  test: [maxMemory: 768, minMemory: 64, debug: false, maxPerm: 256, daemon:true],
+  // configure settings for the run-app JVM
+  run: [maxMemory: 768, minMemory: 64, debug: false, maxPerm: 256, forkReserve:false],
+  // configure settings for the run-war JVM
+  war: [maxMemory: 768, minMemory: 64, debug: false, maxPerm: 256, forkReserve:false],
+  // configure settings for the Console UI JVM
+  console: [maxMemory: 768, minMemory: 64, debug: false, maxPerm: 256]
+]
+
+grails.project.dependency.resolver = "maven" // or ivy
 grails.project.dependency.resolution = {
-    // inherit Grails' default dependencies
-    inherits("global") {
-        // specify dependency exclusions here; for example, uncomment this to disable ehcache:
-        // excludes 'ehcache'
-    }
-    log "error" // log level of Ivy resolver, either 'error', 'warn', 'info', 'debug' or 'verbose'
-    checksums true // Whether to verify checksums on resolve
-    legacyResolve false // whether to do a secondary resolve on plugin installation, not advised and here for backwards compatibility
+  // inherit Grails' default dependencies
+  inherits("global") {
+    // specify dependency exclusions here; for example, uncomment this to disable ehcache:
+    // excludes 'ehcache'
+  }
+  log "error" // log level of Ivy resolver, either 'error', 'warn', 'info', 'debug' or 'verbose'
+  checksums true // Whether to verify checksums on resolve
+  legacyResolve false // whether to do a secondary resolve on plugin installation, not advised and here for backwards compatibility
 
-    repositories {
-        inherits true // Whether to inherit repository definitions from plugins
+  repositories {
+    inherits true // Whether to inherit repository definitions from plugins
 
-        grailsPlugins()
-        grailsHome()
-        grailsCentral()
+    grailsPlugins()
+    grailsHome()
+    mavenLocal()
+    grailsCentral()
+    mavenCentral()
 
-        mavenLocal()
-        mavenCentral()
+    // uncomment these (or add new ones) to enable remote dependency resolution from public Maven repositories
+    //mavenRepo "http://snapshots.repository.codehaus.org"
+    //mavenRepo "http://repository.codehaus.org"
+    //mavenRepo "http://download.java.net/maven/2/"
+    //mavenRepo "http://repository.jboss.com/maven2/"
 
-        // uncomment these (or add new ones) to enable remote dependency resolution from public Maven repositories
-        //mavenRepo "http://snapshots.repository.codehaus.org"
-        //mavenRepo "http://repository.codehaus.org"
-        //mavenRepo "http://download.java.net/maven/2/"
-        //mavenRepo "http://repository.jboss.com/maven2/"
-    }
+    // for spring-security-core
+    mavenRepo 'http://repo.spring.io/milestone'
+  }
 
-    dependencies {
-        // specify dependencies here under either 'build', 'compile', 'runtime', 'test' or 'provided' scopes e.g.
+  dependencies {
+    // specify dependencies here under either 'build', 'compile', 'runtime', 'test' or 'provided' scopes e.g.
 
-        // runtime 'mysql:mysql-connector-java:5.1.22'
-        test "org.spockframework:spock-grails-support:0.7-groovy-2.0"
-    }
+    // runtime 'mysql:mysql-connector-java:5.1.29'
+    // runtime 'org.postgresql:postgresql:9.3-1101-jdbc41'
+    test "org.grails:grails-datastore-test-support:1.0-grails-2.4"
+  }
 
-    plugins {
-        runtime ":hibernate:$grailsVersion"
-        runtime ":jquery:1.8.3"
-        runtime ":resources:1.1.6"
+  plugins {
+    // plugins for the build system only
+    build ":tomcat:7.0.55"
+    // plugins for the compile step
+    compile ":scaffolding:2.1.2"
+    compile ':cache:1.1.7'
 
-        // Uncomment these (or add new ones) to enable additional resources capabilities
-        //runtime ":zipped-resources:1.0"
-        //runtime ":cached-resources:1.0"
-        //runtime ":yui-minify-resources:0.1.5"
+    compile ':spring-security-core:2.0-RC4'
 
-        build ":tomcat:$grailsVersion"
+    // plugins needed at runtime but not for compilation
+    runtime ":hibernate4:4.3.5.5" // or ":hibernate:3.6.10.17"
+    //runtime ":database-migration:1.4.0"
+    runtime ":jquery:1.11.1"
+    runtime ":resources:1.2.8"
 
-        runtime ":database-migration:1.3.2"
+    // Uncomment these (or add new ones) to enable additional resources capabilities
+    //runtime ":zipped-resources:1.0.1"
+    //runtime ":cached-resources:1.1"
+    //runtime ":yui-minify-resources:0.1.5"
 
-        compile ':cache:1.0.1'
-
-        compile ':spring-security-core:1.2.7.3'
-
-        test(":spock:0.7") {
-            exclude "spock-grails-support"
-        }
-    }
+  }
 }
