+++ grails-app/conf/UrlMappings.groovy
@@ -2,14 +2,14 @@ class UrlMappings {
 
 	static mappings = {
 		"/$controller/$action?/$id?"{
-			constraints {
+      constraints {
 				// apply constraints here
 			}
 		}
 
 		"/"(view:"/index")
-        "/login/$action?"(controller: "login")
-        "/logout/$action?"(controller: "logout")
-        "500"(view:'/error')
+    "/login/$action?"(controller: "login")
+    "/logout/$action?"(controller: "logout")
+    "500"(view:'/error')
 	}
 }
