--- grails-app/controllers/com/objectpartners/security/LogoutController.groovy
@@ -1,14 +1,0 @@
-package com.objectpartners.security
-
-import org.codehaus.groovy.grails.plugins.springsecurity.SpringSecurityUtils
-
-class LogoutController {
-
-    /**
-     * Index action. Redirects to the Spring security logout uri.
-     */
-    def index = {
-        // TODO put any pre-logout code here
-        redirect uri: SpringSecurityUtils.securityConfig.logout.filterProcessesUrl // '/j_spring_security_logout'
-    }
-}
